оптимизация общей структуры
