+++ js/main.js
@@ -1,4 +1,2 @@
-import './data.js';
-import './generating-ads.js';
-import './page-state.js';
+
 import './map.js';
