+++ js/map.js
@@ -1,12 +1,11 @@
 /* global L:readonly */
 import { activeState } from './page-state.js';
 import { ads } from './data.js';
-import { homeTypeToReadable } from './common/maps.js';
-import { getAdsPhotos } from './generating-ads.js';
-import { getAdsFeauters } from './generating-ads.js';
+import { createCustomPopup } from './generating-popups.js'
 
 const ROUNDING_COORDINATES = 5;
 
+//Инициализация карты
 const map = L.map('map-canvas')
   .on('load', () => {
     activeState();
@@ -16,7 +15,6 @@ const map = L.map('map-canvas')
     lng: 139.69171,
   }, 12);
 
-
 L.tileLayer(
   'https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png',
   {
@@ -24,12 +22,14 @@ L.tileLayer(
   },
 ).addTo(map);
 
+// Кастомная иконка главной метки
 const mainPinIcon = L.icon({
   iconUrl: '../img/main-pin.svg',
   iconSize: [52, 52],
   iconAnchor: [26, 52],
 });
 
+// Добавление главной метки
 const mainMarker = L.marker(
   {
     lat: 35.6895,
@@ -42,6 +42,7 @@ const mainMarker = L.marker(
 );
 mainMarker.addTo(map);
 
+// Передача координат от главной метки в поле формы
 const addressInput = document.querySelector('#address');
 addressInput.setAttribute('readonly', 'readonly');
 const defaultLatitude = mainMarker._latlng.lat;
@@ -54,23 +55,7 @@ mainMarker.on('move', (evt) => {
   addressInput.value = `${latitude.toFixed(ROUNDING_COORDINATES)}, ${longitude.toFixed(ROUNDING_COORDINATES)}`
 });
 
-const createCustomPopup = (author, offer) => {
-  const balloonTemplate = document.querySelector('#card').content.querySelector('.popup');
-  const popupElement = balloonTemplate.cloneNode(true);
-
-  popupElement.querySelector('.popup__avatar').src = author.avatar;
-  popupElement.querySelector('.popup__title').textContent = offer.title;
-  popupElement.querySelector('.popup__text--address').textContent = offer.adress;
-  popupElement.querySelector('.popup__text--price').textContent = `${offer.price} ₽/ночь`;
-  popupElement.querySelector('.popup__type').textContent = homeTypeToReadable[offer.type];
-  popupElement.querySelector('.popup__text--capacity').textContent = `${offer.rooms} комнаты для ${offer.guests} гостей`;
-  popupElement.querySelector('.popup__text--time').textContent = `Заезд после ${offer.checkin}, выезд до ${offer.checkout}`;
-  getAdsFeauters(offer, popupElement);
-  popupElement.querySelector('.popup__description').textContent = offer.description;
-  getAdsPhotos(offer, popupElement);
-  return popupElement;
-};
-
+// Добавление обычных меток с попапами похожих объявлений
 ads.forEach(({location, offer, author}) => {
   const secondaryPinIcon = L.icon({
     iconUrl: '../img/pin.svg',
